requirements.txt
boto3==1.14.48
botocore==1.17.48
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-bootstrap-pagination==1.7.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-pagination-bootstrap==2.3.3
django-social-share==1.4.0
django-storages==1.9.1
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
docutils==0.15.2
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jsmin==2.2.2
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.1.0
